big refactor: shader weather params
