+++ WickedEngine/shaders/bvh_primitivesCS.hlsl
@@ -76,7 +76,7 @@ void main(uint3 DTid : SV_DispatchThreadID, uint groupIndex : SV_GroupIndex)
 	float3 minAABB = min(P0, min(P1, P2));
 	float3 maxAABB = max(P0, max(P1, P2));
 	float3 centerAABB = (minAABB + maxAABB) * 0.5f;
-	const uint mortoncode = morton3D((centerAABB - g_xFrame.WorldBoundsMin) * g_xFrame.WorldBoundsExtents_rcp);
+	const uint mortoncode = morton3D((centerAABB - GetScene().aabb_min) * GetScene().aabb_extents_rcp);
 	primitiveMortonBuffer[primitiveID] = (float)mortoncode; // convert to float before sorting
 
 }
