+++ WickedEngine/shaders/renderlightmapPS.hlsl
@@ -65,7 +65,7 @@ float4 main(Input input) : SV_TARGET
 					float3 atmosphereTransmittance = 1.0;
 					if (g_xFrame.Options & OPTION_BIT_REALISTIC_SKY)
 					{
-						atmosphereTransmittance = GetAtmosphericLightTransmittance(g_xFrame.Atmosphere, surface.P, L, texture_transmittancelut);
+						atmosphereTransmittance = GetAtmosphericLightTransmittance(GetWeather().atmosphere, surface.P, L, texture_transmittancelut);
 					}
 					
 					float3 lightColor = light.GetColor().rgb * light.GetEnergy() * atmosphereTransmittance;
