+++ WickedEngine/shaders/surfel_raytraceCS.hlsl
@@ -159,7 +159,7 @@ void main(uint3 DTid : SV_DispatchThreadID)
 						float3 atmosphereTransmittance = 1.0;
 						if (g_xFrame.Options & OPTION_BIT_REALISTIC_SKY)
 						{
-							atmosphereTransmittance = GetAtmosphericLightTransmittance(g_xFrame.Atmosphere, surface.P, L, texture_transmittancelut);
+							atmosphereTransmittance = GetAtmosphericLightTransmittance(GetWeather().atmosphere, surface.P, L, texture_transmittancelut);
 						}
 
 						float3 lightColor = light.GetColor().rgb * light.GetEnergy() * atmosphereTransmittance;
